% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitLandS.R
\name{knit_LB}
\alias{knit_LB}
\title{Render R Markdown}
\usage{
knit_LB(input, ...)
}
\arguments{
\item{input}{The input file to be rendered. This can be an R script (.R),
an R Markdown document (.Rmd), or a plain markdown document.}

\item{...}{Additional arguments passed to the rmarkdown::render function.}
}
\value{
When \code{run_pandoc = TRUE}, the compiled document is written into
the output file, and the path of the output file is returned. When
\code{run_pandoc = FALSE}, the path of the Markdown output file, with
attributes \code{knit_meta} (the \pkg{knitr} meta data collected from code
chunks) and \code{intermediates} (the intermediate files/directories
generated by \code{render()}).
}
\description{
Render the input file to a specific output folder. If the Output folder does not exist it will be rendered in the same folder as the input file.
}
\details{
You can also setup the \code{knit} function in the YAML header of the R Markdown file. See the example below.
}
\examples{
\dontrun{
library(rmarkdown)

# Render the default (first) format defined in the file
render("input.Rmd")
}


}
