% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indiedown_pdf_document.R
\name{TemplateLandS}
\alias{TemplateLandS}
\title{Customized R Markdown Document}
\usage{
TemplateLandS(...)
}
\arguments{
\item{...}{Passed on to \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document()}}.}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown:render]{rmarkdown::render()}}.
}
\description{
\code{TemplateLandS()} is the main function of the TemplateLandS package. Use as
\code{output: TemplateLandS::TemplateLandS} in the R Markdown YAML header.
}
